Much faster integration tests
